PEP8 Renames Utilities module
