+++ Future/_KinectFrames.py
@@ -13,7 +13,7 @@ import os;
 import xlwt;
 from math import pi,tan;
 
-import TopMouseTracker.Utilities as utils;
+import TopMouseTracker.utilities as utils;
 
 
 class Kinect() :
