+++ _RunOffline.py
@@ -25,7 +25,7 @@ if os.getcwd() !=  _topMouseTrackerDir : #If the current working directory is no
 # Loads TMT modules#
 
 import TopMouseTracker.Parameters as params
-import TopMouseTracker.Utilities as utils
+import TopMouseTracker.utilities as utils
 import TopMouseTracker.IO as IO
 import TopMouseTracker._Tracker as tracker
 import TopMouseTracker.Analysis as analysis
