+++ _Tracker.py
@@ -24,7 +24,7 @@ from email.mime.multipart import MIMEMultipart;
 from PIL import Image;
 
 import TopMouseTracker.Parameters as params;
-import TopMouseTracker.Utilities as utils;
+import TopMouseTracker.utilities as utils;
 import TopMouseTracker.IO as IO;
 
 class TopMouseTracker():
